Simplify terrafying
Terraform now does all the things that terrafying tried to do for
you beyond the JSON generation, so lets delete all those unneeded
features
