+++ lib/terrafying/version.rb
@@ -1,4 +1,3 @@
 module Terrafying
   VERSION = "0.0.0" # will be inserted by Drone
-  CLI_VERSION = "0.11.3"
 end
